Remove compiler warning.
git-svn-id: https://aeon.stsci.edu/ssb/svn/astrolib/trunk/pywcs@1442 90a0a646-be8a-0410-bb88-9290da87bc01

Former-commit-id: 42e6cc7e60904d08952573121524ae6c9bbea327

+++ src/pyutil.c
@@ -269,7 +269,7 @@ shape_to_string(
 
   str[0] = 0;
   for (i = 0; i < ndims; ++i) {
-    snprintf(value, 32, "%d", dims[i]);
+      snprintf(value, 32, "%d", (int)dims[i]);
     strncat(str, value, 32);
     if (i != ndims - 1) {
       strncat(str, "x", 2);
